// src/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface HpBannerAttributes {
  title: string
  slug: string
  status:string
  description?: string | null
  cta_text: string
  background_image?: Media | number | null
}